Refactor relationshipDropdown by filtering
Rather than building up a list of gender codes for each combination
of clients, this attempts to filter out relationship types based on
whether the clients are compatible with them, one step at a time.
This makes each individual step a lot simpler and drastically cuts
down on the overal code footprint.